Fix Typo and Add Custom Probes interface
